+++ docs/release-notes.md
@@ -2,6 +2,10 @@
 
 # Release notes
 
+## Upcoming release
+
+* Fixed parameterless constructors not being utilized.
+
 ## 2.0.4
 Released 25 August 2024.